add bottom_sheet and merge file
