+++ src/component/user/dataBottomPopUp.js
@@ -1,0 +1,84 @@
+export const dataPopUp = [
+  {
+    id: 1,
+    icon: 'settings-outline',
+    text: 'Cài đặt',
+  },
+  {
+    id: 2,
+    icon: 'ios-timer-outline',
+    text: 'Hoạt động của bạn',
+  },
+  {
+    id: 3,
+    icon: 'md-time-outline',
+    text: 'Kho lưu trữ',
+  },
+  {
+    id: 4,
+    icon: 'qr-code-outline',
+    text: 'Mã QR',
+  },
+  {
+    id: 5,
+    icon: 'bookmark-outline',
+    text: 'Đã lưu',
+  },
+  {
+    id: 6,
+    icon: 'card-outline',
+    text: 'Đơn đặt hàng và thanh toán',
+  },
+  {
+    id: 7,
+    icon: 'menu-outline',
+    text: 'Bản thân',
+  },
+  {
+    id: 8,
+    icon: 'md-star-outline',
+    text: 'Yêu thích',
+  },
+  {
+    id: 9,
+    icon: 'person-add-outline',
+    text: 'Khám phá mọi người',
+  },
+  {
+    id: 10,
+    icon: 'information-circle-outline',
+    text: 'Cập nhật tính năng tin nhắn',
+  },
+];
+export const dataPopUpCreate = [
+  {
+    id: 1,
+    icon: 'movie-settings-outline',
+    text: 'Thước phim',
+  },
+  {
+    id: 2,
+    icon: 'table-large',
+    text: 'Bài Viết',
+  },
+  {
+    id: 3,
+    icon: 'play-speed',
+    text: 'Tin',
+  },
+  {
+    id: 4,
+    icon: 'heart-circle-outline',
+    text: 'Tin nổi bật',
+  },
+  {
+    id: 5,
+    icon: 'access-point-network',
+    text: 'Phát trực tiếp',
+  },
+  {
+    id: 6,
+    icon: 'book-account-outline',
+    text: 'Hướng dẫn',
+  },
+];
