+++ src/style/search/styleImage.js
